Reorganise tests into separate modules
